refactor: 변경된 예약 정보 도메인 파일 적용
