+++ app/src/main/java/com/example/team25/ui/main/status/interfaces/OnCheckReportClickListener.kt
@@ -1,6 +1,6 @@
 package com.example.team25.ui.main.status.interfaces
 
-import com.example.team25.ui.main.status.data.ReservationInfo
+import com.example.team25.domain.model.ReservationInfo
 
 interface OnCheckReportClickListener {
     fun onCheckReportClicked(item: ReservationInfo)
