+++ app/src/main/java/com/example/team25/ui/main/status/interfaces/OnRequestCancelClickListener.kt
@@ -1,6 +1,6 @@
 package com.example.team25.ui.main.status.interfaces
 
-import com.example.team25.ui.main.status.data.ReservationInfo
+import com.example.team25.domain.model.ReservationInfo
 
 interface OnRequestCancelClickListener {
     fun onRequestCancelClicked(item: ReservationInfo)
